add switch apparence
