+++ .eslintrc.js
@@ -2,7 +2,7 @@ module.exports = {
     root: true,
     parserOptions: {
         sourceType: `module`,
-        ecmaVersion: 2020,
+        ecmaVersion: 2022,
         ecmaFeatures: {
             globalReturn: false,
             impliedStrict: false,
