+++ jest.config.js
@@ -4,4 +4,4 @@ module.exports = {
     transform: {
         "^.+\\.js$": `babel-jest`
     }
-};
+}
