+++ package.json
@@ -29,7 +29,7 @@
   "devDependencies": {
     "@babel/core": "^7.21.4",
     "@babel/preset-env": "^7.21.4",
-    "eslint": "^7.32.0",
+    "eslint": "^8.50.0",
     "eslint-plugin-jest": "^27.2.1",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
